+++ test/schemas/Object.test.js
@@ -9,4 +9,9 @@ describe(schema.Object.name, () => {
     });
     expect(normalize({ user: { id: 1 } }, object)).toMatchSnapshot();
   });
+
+  it(`auto-converts plain objects to ${schema.Object.name}`, () => {
+    const userSchema = new schema.Entity('user');
+    expect(normalize({ user: { id: 1 } }, { user: userSchema })).toMatchSnapshot();
+  });
 });