+++ CHANGELOG.md
@@ -6,13 +6,14 @@
 * `schema.Entity`
     * `processStrategy` for modifying `Entity` objects before they're moved to the `entities` stack. [example](/#TODO)
     * `mergeStrategy` for merging with multiple entities with the same ID. [example](/#TODO)
+* Added `schema.Object`, with a shorthand of `{}`
+* Added `schema.Array`, with a shorthand of `[ schema ]`
 
 ## Changed
 
-* `arrayOf` has been replaced by `schema.Array`
+* `arrayOf` has been replaced by `schema.Array` or `[]`
 * `unionOf` has been replaced by `schema.Union`
 * `valuesOf` has been replaced by `schema.Values`
-* `{}` as a part of a schema definition has been replaced by `schema.Object`
 
 ## Removed